pyproject.toml
[project]
name = "timetable-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf2>=2.8.5",
    "ics>=0.7.2",
    "nicegui>=3.4.0",
    "pandas>=2.3.3",
    "pdfplumber>=0.11.8",
]
